Abstracted validations into a shared concern to make available to MailRequest object
